[UPDATE]HeaderOnly is not a good design and changed to be used as a DLL.
