+++ WinAppCore/WinAppCore/WinAppCore.vcxproj.filters
@@ -27,9 +27,6 @@
     <ClCompile Include="src\console_output.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="src\config.cpp">
-      <Filter>ソース ファイル</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\config.h">
