+++ WinAppCore/WinAppCore/include/config.h
@@ -7,25 +7,3 @@
 #else
 #define WIN_APP_CORE __declspec(dllimport)
 #endif
-
-#include <string>
-
-namespace WACore
-{
-
-class WIN_APP_CORE CoutConfig
-{
-public:
-    static std::string& START_TAG();
-    static std::string& END_TAG();
-
-    static std::string& COLOR_NORMAL();
-    static std::string& COLOR_ERROR();
-    static std::string& COLOR_WARNING();
-    static std::string& COLOR_INFO();
-    static std::string& COLOR_DEBUG();
-
-    static std::string& COLOR_RESET();
-};
-
-} // namespace WACore
