+++ WinAppCore/WinAppCore/include/console_output.h
@@ -9,6 +9,21 @@
 namespace WACore
 {
 
+class WIN_APP_CORE CoutConfig
+{
+public:
+    static std::string& START_TAG();
+    static std::string& END_TAG();
+
+    static std::string& COLOR_NORMAL();
+    static std::string& COLOR_ERROR();
+    static std::string& COLOR_WARNING();
+    static std::string& COLOR_INFO();
+    static std::string& COLOR_DEBUG();
+
+    static std::string& COLOR_RESET();
+};
+
 WIN_APP_CORE unsigned int& ErrorCount();
 WIN_APP_CORE unsigned int& WarningCount();
